+++ bumpr/__init__.py
@@ -17,14 +17,20 @@ __description__ = 'Version bumper and Python package releaser'
 
 
 def main():
+    import sys
     from bumpr import log
     log.init()
 
     from bumpr.config import Config
     from bumpr.releaser import Releaser
+    from bumpr.helpers import BumprError
     from logging import DEBUG, INFO, getLogger
 
     config = Config.parse_args()
     getLogger().setLevel(DEBUG if config.verbose else INFO)
-    releaser = Releaser(config)
-    releaser.release()
+    try:
+        releaser = Releaser(config)
+        releaser.release()
+    except BumprError as error:
+        getLogger(__name__).error(error.message)
+        sys.exit(1)