+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Current
 - Rely on VCS for tracking files and ensure working copy is clean
 - Added option ``--nocommit``
 - Ensure dry run does not write or execute anything
-- Better output
+- Better output and error handling
 - Group parameters in help
 
 0.1.0 (2013-08-22)